# log: Simplified and optimized the structlog