--- a/cu29/src/curuntime.rs
+++ b/cu29/src/curuntime.rs
@@ -285,6 +285,7 @@ mod tests {
         ))
     }
 
+    #[derive(Debug)]
     struct FakeWriter {}
 
     impl<E: Encode> WriteStream<E> for FakeWriter {