--- a/cu29_export/build.rs
+++ b/cu29_export/build.rs
@@ -1,5 +1,5 @@
 use std::env;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 
 fn main() {