--- a/examples/cu_standalone_structlog/build.rs
+++ b/examples/cu_standalone_structlog/build.rs
@@ -1,0 +1,5 @@
+use std::env;
+fn main() {
+    let out_dir = env::var("OUT_DIR").unwrap();
+    println!("cargo:rustc-env=OUT_DIR={}", out_dir);
+}