Notable change: I completely removed the queue in the runtime as the overhead was larger than a simple unbuffered mutex. 

I also build a quick benchmark to compare the speed with the more traditional Rust text logging.

```
sal ➜  cu_standalone_structlog (gbin/structlog_optim) ✗ source back2back.sh
   Compiling cu_standalone_structlog v0.2.1 (/home/gbin/projects/copper/copper-rs/examples/cu_standalone_structlog)
    Finished `release` profile [optimized] target(s) in 1.92s
     Running `/home/gbin/projects/copper/copper-rs/target/release/standardlog_perf`
Total time: 34.220 s in ./logfile.txt
.rw-r--r-- gbin gbin 101 MB Sat Aug 10 10:45:55 2024  logfile.txt
   Compiling cu_standalone_structlog v0.2.1 (/home/gbin/projects/copper/copper-rs/examples/cu_standalone_structlog)
CLog: ==================================================================================
CLog: Interned strings are stored in: "/home/gbin/projects/copper/copper-rs/target/release/cu29_log_index"
CLog:    [r] is reused index and [n] is new index.
CLog: ==================================================================================
CLog: #001 [r] -> This is the logline {} associated with the log Logging and some more = {}, {}.
    Finished `release` profile [optimized] target(s) in 1.84s
     Running `/home/gbin/projects/copper/copper-rs/target/release/structlog_perf`
Total time: 70.342 ms in ./logfile.bin
.rw-r--r-- gbin gbin 28 MB Sat Aug 10 10:45:57 2024  logfile.bin
```

60% reduction in logged size and 486x faster walltime